% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{attr}
\alias{attr}
\alias{\%.\%}
\alias{\%.\%<-}
\title{Interface to read and write attributes.}
\usage{
x \%.\% which

\%.\%(x, which) <- value
}
\arguments{
\item{x}{An object whose attributes are to be accessed.}

\item{which}{A non-empty string specifying which attribute is to be accessed.}

\item{value}{The new value of the attribute, or \code{NULL} to remove the
attribute.}
}
\description{
Interface to easily read and write attributes.
}
\examples{
# Create a 2 by 5 matrix:
x <- 1:10
x\%.\%dim <- c(2, 5)
# The previous line is identical to: attr(x, "dim") <- c(2, 5)

}
\seealso{
\code{\link{attr}}, \code{\link{attributes}}
}
