% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{bigRR.data}
\alias{bigRR.data}
\alias{bigRR.data-class}
\title{The \code{bigRR.data} class

This class contains results of running the \code{\link[cgmisc]{gwaa2bigRR}} function.}
\description{
The \code{bigRR.data} class

This class contains results of running the \code{\link[cgmisc]{gwaa2bigRR}} function.
}
\section{Slots}{

 \describe{
   \item{\code{y}:}{a \code{vector} with response variable,}
   \item{\code{X}:}{design \code{matrix} for fixed effects,}
   \item{\code{Z}:}{design \code{matrix} for random effects.}
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@slu.se}>
}

