% Generated by roxygen2 (4.0.2): do not edit by hand
\name{boxplot.snp}
\alias{boxplot.snp}
\title{Box plot with genotypes}
\usage{
\method{boxplot}{snp}(GWASdata, marker, trait, recode = F, ...)
}
\arguments{
\item{GWASdata}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}}

\item{marker}{SNP name}

\item{trait}{trait name}

\item{recode}{if genotypes are to be recoded to 'AA','Aa','aa'.}
}
\value{
null
}
\description{
Makes boxplots of the individuals in every genotypic class, based on a SNP.
Works for both outbreed (three boxes) and inbreed (two boxes) data.
}
\author{
Simon Forsberg <\email{simon.forsberg@slu.se}>
}

