% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxplot.snp.twoWay.R
\name{boxplot.snp.twoWay}
\alias{boxplot.snp.twoWay}
\title{SNP_box_twoWay}
\usage{
boxplot.snp.twoWay(data, marker1, marker2, trait, legend = F, ...)
}
\arguments{
\item{data}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}}

\item{marker1}{The name or index of one marker/SNP in data}

\item{marker2}{The name or index of a second marker/SNP in data}

\item{trait}{The name of a phenotype in data}

\item{legend}{If TRUE, a legend showing the alternative genotypes for the two SNPs will be added to the plot.}
}
\value{
null
}
\description{
Same as boxplot.snp but based on two SNPs.
This function makes box plots, showing the Genotype x Phenotype map for two loci.
Works for both outbreed (nine boxes) and inbreed (four boxes) data.
Can be used to look for epistasis.
}
\author{
Simon Forsberg  <\email{simon.forsberg@slu.se}>
}

