\name{cgmisc}
\alias{cgmisc-package}
\alias{cgmisc}
\docType{package}
\title{Various functions useful in computational genetics}
\description{
Package contains miscellaneous functions useful in computational genetics, especially in genome-wide association studies (GWAS).
}
\details{
\tabular{ll}{
Package: \tab cgmisc\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2013-04-09\cr
License: \tab GPL (>=2.10)\cr
}
Currently the package contains the following functions: \cr
- plot.manhattan.LD -- a function for plotting LD pattern on Manhattan plot, \cr
- compute.Fstats -- a function for computing Fst for two populations, \cr
- clump.markers -- an R implementation of clumping algorithm known from PLINK, \cr
- plot.clumps -- a function for plotting clumping results, \cr
}
\author{
Author(s): Marcin Kierczak

Maintainer: Marcin Kierczak <marcin.kierczak@slu.se>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\keyword{ computational genetics }
\keyword{ genetics }
\seealso{
\code{\link[<GenABEL>:<GenABEL>-package]{<GenABEL>}}
}
\examples{
}
