% Generated by roxygen2 (4.0.2): do not edit by hand
\name{choose.top.snps}
\alias{choose.top.snps}
\title{Select the highest r2 SNPs to a given top marker.}
\usage{
choose.top.snps(data, chr, region, index.snp)
}
\arguments{
\item{data}{data object in \code{\link[GenABEL]{gwaa.data-class}}}

\item{chr}{chromosome number (name) to be displayed}

\item{region}{a vector of two coordinates}

\item{index.snp}{the reference SNP (name of the marker)}
}
\value{
matrix with highest r2 SNPs with their r2 values and coordinates
}
\description{
Calculates r2 and selects SNPs with highest r2 to a given top marker.
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}> Jagoda Jablonska <\email{jagoda100jablonska@gmail.com}
}

