% Generated by roxygen2 (4.0.2): do not edit by hand
\name{chr.x.fix.canfam}
\alias{chr.x.fix.canfam}
\title{Fix canine X chromosome naming}
\usage{
chr.x.fix.canfam(data, assembly = "canFam3")
}
\arguments{
\item{data}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}}
}
\value{
a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}

an object of \code{\link[GenABEL]{gwaa.data-class}} class
}
\description{
Function is renaming canine sex chromosome from "39" to "X".
Pseudoautosomal (PAR) region (340,578 bp to 6,642,728 bp canFam3 and 314,872 bp to 6,598,016 bp canFam2) on the
canine X chromosome is renamed to "parX"
}
\details{
Currently supported assemblies are: "canFam3","canFam2"
}
\examples{
\dontrun{
fixed.data <- Xfix.canfam(data, assembly='canFam3')
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>, Matteo Bianchi <\email{Matteo.Bianchi@imbim.uu.se}>
}
\seealso{
\code{\link[GenABEL]{gwaa.data-class}}
}
\keyword{PAR}
\keyword{X}
\keyword{canine}
\keyword{chromosome}
\keyword{pseudoautosomal}
\keyword{sex}

