% Generated by roxygen2 (4.0.2): do not edit by hand
\name{create.haploview.info}
\alias{create.haploview.info}
\title{Prepare a PHASE input file from gwaa data}
\usage{
create.haploview.info(data, chr, coords, outFile)
}
\arguments{
\item{data}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}},}

\item{chr}{chromosome number (name) to be used,}

\item{coords}{a vector of two coordinates that define the region. If not set, the entire chromosome will be used,}

\item{outFile}{a complete path and name of output file.}
}
\description{
Function for preparing PHASE and fastPHASE input files from \code{\link[GenABEL]{gwaa.data-class}} object.
It can be used on a single entire chromosome or on a specified chromosomal region.
}
\examples{
\dontrun{
create.Haploview.info(data=data.qc2, chr=2, coords=c(3030587,5030587), outFile="~user/test.info")
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\seealso{
\code{\link[GenABEL]{gwaa.data-class}}
}
\keyword{PHASE}
\keyword{fastPHASE}
\keyword{genetics}
\keyword{haplotype}

