% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/epistasis.scan.r
\name{epistasis.scan}
\alias{epistasis.scan}
\title{epistasisScan}
\usage{
epistasis.scan(data, SNP, trait.name = NULL, pheno = NULL)
}
\arguments{
\item{data}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}}

\item{SNP}{The name or index of one marker/SNP in data}

\item{trait.name}{The name of a phenotype in data}

\item{pheno}{A vector with phenotypes}
}
\value{
a list with all the models (lm-objects),
the p-values of all models and coefficients, and the genomic coordinates of all scanned SNPs.
}
\description{
Scans for epistasis between a specified SNP and all others.
}
\details{
The function fits a linear model, including interaction term, between SNP and all other markers in data.
}
\author{
Simon Forsberg  <\email{simon.forsberg@slu.se}>
}

