% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Format real numbers for pretty printing.}
\usage{
fmt(digits = 3)
}
\arguments{
\item{digits}{How many digits to be kept?}
}
\description{
Closure to format numeric numbers for pretty printing.
}
