% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fstats.result.r
\docType{class}
\name{fstats.result}
\alias{fstats.result}
\alias{fstats.result-class}
\title{The \code{fstats.result} class

This class contains results returned by the \code{\link[cgmisc]{compute.Fstats}} function.}
\description{
The \code{fstats.result} class

This class contains results returned by the \code{\link[cgmisc]{compute.Fstats}} function.
}
\section{Slots}{

 \describe{
   \item{\code{pops}:}{A \code{list} of per-population statistics}
   \item{\code{glob}:}{A \code{data.frame} of global statistics}
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}

