% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get.adjacent.markers.r
\name{get.adjacent.markers}
\alias{get.adjacent.markers}
\title{Get adjacent markers within a distance around the given marker}
\usage{
get.adjacent.markers(data, marker, size.bp = 1000)
}
\arguments{
\item{data}{a gwaa.data class object as used by the \code{\link[GenABEL]{gwaa.data-class}},}

\item{marker}{a name of the central marker,}

\item{size.bp}{window size in basepairs,}
}
\value{
a matrix of marker adjacency
}
\description{
Returns markers adjacent to a given marker (within a preset distance).
}
\examples{
\dontrun{
adj.markers  <- get.adjacent.markers(data, 'BICF2P647127')
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{adjacent}
\keyword{markers}

