% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.chr.midpoints}
\alias{get.chr.midpoints}
\title{Get middle point for each chromosome}
\usage{
get.chr.midpoints(data)
}
\arguments{
\item{data}{a \code{\link[GenABEL]{gwaa.data-class}} object produced by GenABEL}
}
\value{
a vector of mid-point coordinates

vector of chromosomes midpoints
}
\description{
Returns a vector with chromosome mid-point coordinates
}
\details{
Function result is independent of coordinates used. Function is primarily used
for plotting chromosome labels on the x-axis.
}
\examples{
\dontrun{
 midpoints <- get.chr.midpoints(data)
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{chromosome,}
\keyword{midpoint}

