% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.overlapping.windows}
\alias{get.overlapping.windows}
\title{Get overlapping windows}
\usage{
get.overlapping.windows(data, chr, size = 1250000, overlap = 2500)
}
\arguments{
\item{data}{data object in \code{\link[GenABEL]{gwaa.data-class}},}

\item{chr}{chromosome to be divided into windows - if not specified, the entire genome is taken into consideration,}

\item{size}{size of the window,}

\item{overlap}{size of the overlap.}
}
\value{
List containing windows coordinates and LW matrix with windows and markers.
}
\description{
Divides chromosome into a number of overlapping windows.
Windows and overlap size is defined in terms of a fixed number of base pairs.
One may use returned windows for further analyses.
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\seealso{
\code{\link[cgmisc]{het.for.overlap.wind}}
}
\keyword{overlapping,}
\keyword{windows}

