% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gwaa.to.bed}
\alias{gwaa.to.bed}
\title{Create a BED file with GWAS p-values in a region}
\usage{
gwaa.to.bed(chr, range, gwas, fname = "output.bed")
}
\arguments{
\item{chr}{chromosome to be considered}

\item{range}{a vector of genomic ranges}

\item{gwas}{a \code{\link[GenABEL]{scan.gwaa-class}} object produced by GenABEL}

\item{fname}{name of the output file. By default: output.bed}
}
\value{
a bed file with -log10 of corrected p-value ("Pc1df")
}
\description{
Returns a file containing GWAS p-values for a region in bed format. It can be further used as
UCSC Genome Browser custom track.
}
\examples{
\dontrun{
 chr <- 36
 range <- c(14e6,24e6)
 gwas <- data.qc1
 gwas.region.bed(chr, range, gwas, fname="my_output.bed")
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{bed,}
\keyword{gwas,}
\keyword{p-values}

