% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwaa.to.phase.r
\name{gwaa.to.phase}
\alias{gwaa.to.PHASE}
\alias{gwaa.to.phase}
\alias{gwaa2PHASE}
\alias{gwaa2phase}
\title{Prepare a PHASE input file from gwaa data.}
\usage{
gwaa.to.phase(data, chr, coords = NULL, outFile)
}
\arguments{
\item{data}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}},}

\item{chr}{chromosome nubmber (name) to be used,}

\item{coords}{a vector of coordinates that define the region. If not set, the entire chromosome will be used,}

\item{outFile}{a complete path and name of the desired output file.}
}
\description{
Function for preparing PHASE and fastPHASE input files from \code{\link[GenABEL]{gwaa.data-class}} object.
It can be used on a single entire chromosome or on a specified chromosomal region.
}
\examples{
\dontrun{
gwaa2PHASE(data=data.qc2, chr=2, coords=c(3030587,5030587), outFile="~user/test.phase")
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\seealso{
\code{\link[GenABEL]{gwaa.data-class}}
}
\keyword{PHASE}
\keyword{fastPHASE}
\keyword{genetics}
\keyword{haplotype}

