% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwaa2bigRR}
\alias{gwaa2bigRR}
\title{Converts gwaa-data-class GenABEL object to a bigRR-readable format.}
\usage{
gwaa2bigRR(formula, data, trait)
}
\arguments{
\item{formula}{the part of formula for fixed effects (the ones with no shrinkage)}

\item{data}{data object in \code{\link[GenABEL]{gwaa.data-class}},}

\item{trait}{trait name}
}
\value{
an object of a bigRR-data.class
}
\description{
\code{gwaa2bigRR} converts gwaa-data-class GenABEL object to a bigRR-readable format.
}
\examples{
\dontrun{
   bigRR.data <- gwaa2bigRR(~sex, data.qc0, gt)
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@slu.se}>
}
\references{
\url{}
}

