% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gwaa2vgwas}
\alias{gwaa2vgwas}
\title{Convert GenABEL data structure to vGWAS data structure}
\usage{
gwaa2vgwas(data, trait.name)
}
\arguments{
\item{data}{an object of the \code{\link[GenABEL]{gwaa.data-class}}}

\item{trait.name}{name of the trait of interest}
}
\value{
a vGWAS readable list
}
\description{
Converts  GenABEL data (\code{\link[GenABEL]{gwaa.data-class}} object) to \code{\link[vGWAS]{vGWAS}} input format.
}
\examples{
\dontrun{vGWAS.input.data <- gwaa2vgwas(data=cgmiscdat1, "trait1")}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@slu.se}>
}
\seealso{
\code{\link[GenABEL]{GenABEL}}, \code{\link[vGWAS]{vGWAS}}
}
\keyword{GenABEL,}
\keyword{convert,}
\keyword{gwaa}
\keyword{vGWAS,}

