% Generated by roxygen2 (4.0.2): do not edit by hand
\name{het.overlap.wind}
\alias{het.overlap.wind}
\title{Calculates heterozygosity for windows}
\usage{
het.overlap.wind(data, LW, progress = TRUE)
}
\arguments{
\item{data}{data object in \code{\link[GenABEL]{gwaa.data-class}}}

\item{LW}{a list of windows and coordinates produced by \code{get.overlapping.windows} function.}

\item{progress}{logical if the progress bar is to be displayed}
}
\value{
Matrix of average heterozygosities for every window
}
\description{
Calculates average heterozygosity for overlapping windows
produced by \code{get.overlapping.windows} function
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>, Jagoda Jablonska <\email{jagoda100jablonska@gmail.com}>
}
\seealso{
\code{\link[cgmisc]{get.overlapping.windows}}
}
\keyword{heterozygosity,}
\keyword{window}

