% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{in_between}
\alias{in_between}
\alias{\%in_between\%}
\title{Is a value within an interval?}
\usage{
in_between(x, bounds, inclusive = TRUE)

x \%in_between\% bounds
}
\arguments{
\item{x}{Vector of values to be tested.}

\item{bounds}{Data frame with two columns or vector with two values.}

\item{inclusive}{Are the bounds included within the interval.}
}
\description{
Is a value within an interval?
}
\examples{
2 \%in_between\% c(1, 3)

}
