% Generated by roxygen2 (4.0.1): do not edit by hand
\name{open.region.UCSC}
\alias{open.region.UCSC}
\title{Open UCSC GenomeBrowser for selected region}
\usage{
\method{open}{region.UCSC}(chr, coords, assembly = "canFam3")
}
\arguments{
\item{chr}{number of chromosome you want to see}

\item{coords}{a vector of two values, if the argument is missing the entire chromosome will be taken into consideration}

\item{assembly}{type of assembly (camFam2 or camFam3)}
}
\value{
null
}
\description{
\code{open.region.UCSC} redirects you to the UCSC GenomeBrowser and shows the specified region
a \code{\link[GenABEL]{gwaa.data-class}} object.
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@slu.se}> , Jagoda Jablonska <\email{jagoda100jablonska@gmail.com}>
}

