% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{out}
\alias{out}
\alias{\%out\%}
\alias{\%!in\%}
\title{Value matching (extra)}
\usage{
x \%out\% table

x \%!in\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link{Long vectors} are not supported.}
}
\description{
Value matching (extra)
}
