% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{pac.result}
\alias{pac.result}
\alias{pac.result-class}
\title{The \code{pac.result} class

This class contains results of running the \code{\link[cgmisc]{pop.allele.counts}} function.}
\description{
The \code{pac.result} class

This class contains results of running the \code{\link[cgmisc]{pop.allele.counts}} function.
}
\section{Slots}{

 \describe{
   \item{\code{chr}:}{A \code{vector} of chromosome numbers.}
   \item{\code{map}:}{A \code{vector} of genomic positions.}
   \item{\code{raf.pop1}:}{Reference allele frequency in population 1.}
   \item{\code{raf.pop2}:}{Reference allele frequency in in population 2.}
   \item{\code{n.pop1}:}{Number of individuals in population 1.}
   \item{\code{n.pop2}:}{Number of individuals in population 2.}
   \item{\code{p.values}:}{A \code{vector} of p-values.}
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}

