% Generated by roxygen2 (4.0.1): do not edit by hand
\name{phase2haploview}
\alias{phase2haploview}
\title{Converts PHASE output to HaploView readable format.}
\usage{
phase2haploview(input, output = NULL)
}
\arguments{
\item{input}{a *.out file containing output from PHASE,}

\item{output}{filename for the result file (optional).}
}
\description{
Converts PHASE output to HaploView-friendly format.
}
\examples{
\dontrun{
 con <- "~/Research/Behavior/results/2014-02-11_mh_aggr_chr36.phase.out"
 phase2haploview(input=con)
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{HaploView,}
\keyword{PHASE,}
\keyword{convert}

