% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.erv}
\alias{plot.erv}
\title{Plot endogenous retroviral sequences in canine genome.}
\usage{
\method{plot}{erv}(chr = NA, coords = c(NA, NA), src = "canFam3cgmisc.db",
  scale = c(0.2, 0.1))
}
\arguments{
\item{chr}{chromosome, e.g. "chr17",}

\item{coords}{a vector of coordinates,}

\item{src}{source database, currently only canFam3 is supported,}

\item{scale}{scaling factors for x and y axis.}
}
\value{
null
}
\description{
Plot endogenous retroviral sequences (ERV) in canine genome.
}
\details{
Plots -log10(score) on the y axis (the higher the score is, the more probable that the hit represents an actual ERV)
and coordinates in Mbp on the x axis. Width of the bars covers entire ERV.
}
\examples{
\dontrun{plot.erv("chr16", coords=c(1e6,2e6))
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>, Patric Jern <\email{Patric.Jern@imbim.uu.se}>
}
\keyword{ERV}
\keyword{plot,}

