% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.genes.R
\name{plot.genes}
\alias{plot.genes}
\title{Plot genes provided in a bed file}
\usage{
plot.genes(region, chr, bed.path = NULL)
}
\arguments{
\item{region}{\code{\link[GenomicRanges]{IRanges}}}

\item{bed.path}{is a path to a bed file containing gene information.}
}
\value{
ggplot2 plot
}
\description{
Using information provided in a bed file, plot genes, together with their names
and strand information.
}
\details{
When using UCSC-provided bigBed annotation files you need to first use the
bigBedToBed tool (available from: http://hgdownload.cse.ucsc.edu/admin/exe/YOUR_OS_DIRECTORY) to
get the right format. At this point you can also extract the desired chromosome.
Then you need to strip the last column of the bed file using, for instance awk:
awk '{out=""; for(i=1;i<=15;i++){out=out" "$i}; print out}' my.bed > myFinal.bed
Here, we provide a Broad Improved Canine Annotation v.1 protein coding genes for the canine chromosome 2
(see: https://www.broadinstitute.org/ftp/pub/vgb/dog/trackHub/hub.txt).
}
\author{
Veronika Scholz <\email{veronikascholz@gmail.com}>
}

