% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.LD.decay}
\alias{plot.LD.decay}
\title{Plots LD decay pattern for a given data}
\usage{
\method{plot}{LD.decay}(data, N = 200, dmin = NA, dmax = NA, ...)
}
\arguments{
\item{data}{a  \code{\link[GenABEL]{gwaa.data-class}} object}

\item{N}{number of bins used to partition the distance between markers}

\item{dmin}{minimal distance to consider}

\item{maximal}{distance to consider}

\item{...}{the remaining parameters that pass to the plot function}
}
\description{
Function plots the LD decay with distance between markers for a given data.
LD is measured using r2 statistics.
}
\examples{
\dontrun{
dat <- dat[,dat@gtdata@chromosome == 3]
plot.LD.decay(data=dat, N=200, dmin=0, dmax=1e6, main="Chr3")
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{LD}
\keyword{LD,}
\keyword{decay,}
\keyword{plot,}
\keyword{r2}

