% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.manhattan.LD}
\alias{plot.manhattan.LD}
\title{Plot LD pattern and MAF in a Manhattan plot}
\usage{
\method{plot}{manhattan.LD}(data, gwas.result, chr, region, index.snp,
  p.value = 0.05, bonferroni = T, mafThreshold = 0.05,
  legend.pos = "def.left")
}
\arguments{
\item{data}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}}

\item{gwas.result}{a scan.gwaa class object as used by \code{\link[GenABEL]{scan.gwaa-class}}}

\item{chr}{chromosome number (name) to be displayed}

\item{region}{a vector of two coordinates to display}

\item{index.snp}{index marker of the reference SNP (name of the marker)}

\item{p.value}{p-value threshold to visualize using red dashed line}

\item{bonferroni}{logical indicating whether Bonferroni-correction shall be used for the displayed p-value threshold}

\item{mafThreshold}{threshold value for minor allele frequency (MAF) -- displayed as a red line on the lower panel}

\item{legend.pos}{a string specifying position of the legend on the plot: "def.left", "def.right" or any other accepted by \code{\link{legend}}}
}
\description{
Function for plotting local LD pattern (relative to a pre-selected marker) on a
Manhattan plot resulting from genome-wide association study (GWAS). Each marker on the plot is
colored according to its LD with the reference marker. Color codes are not continuous, but LD is
discretized into LD intervals. This is useful for examining signals found in a GWAS study.
In addition, minor allele frequency will be plotted in the lower panel.
}
\examples{
\dontrun{
plot.manhattan.LD(data=data.mh1.qc1, gwas.result=mm.mh1, chr=36, region=c(16e6,21e6),
                  index.snp="BICF2P12960", bonferroni=F, legend.pos="default")
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\seealso{
\code{\link[GenABEL]{gwaa.data-class}}, \code{\link[GenABEL]{scan.gwaa-class}}
}
\keyword{LD}
\keyword{Manhattan}
\keyword{disequilibrium}
\keyword{genetics}
\keyword{linkage}

