% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.overlap.R
\name{plot.overlap}
\alias{plot.overlap}
\title{Plots average heterozygosities for overlapping windows}
\usage{
plot.overlap(LW, heter.zyg, ...)
}
\arguments{
\item{LW}{a list of windows and coordinates returned by \code{get.overlap.windows} function}

\item{heter.zyg}{a matrix of heterozygosities returned by \code{\link[cgmisc]{het.overlap.wind}}}
}
\value{
null
}
\description{
Plots average heterozygosities for overlapping windows and the distribution of heterozygosity.
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>, Jagoda Jablonska <\email{jagoda100jablonska@gmail.com}>
}
\seealso{
\code{\link[cgmisc]{get.overlap.windows}}, \code{\link[cgmisc]{het.overlap.wind}}
}
\keyword{heterozygosity,}
\keyword{window}

