% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.overlapping}
\alias{plot.overlapping}
\title{Plots average heterozygosities for overlapping windows}
\usage{
\method{plot}{overlapping}(LW, heter.zyg)
}
\arguments{
\item{LW}{a list of windows and coordinates returned by \code{get.overlapping.windows} function}

\item{heter.zyg}{a matrix of heterozygosities returned by \code{\link[cgmisc]{het.for.overlap.wind}}}
}
\value{
null
}
\description{
Plots average heterozygosities for overlapping windows and the distribution of heterozygosity.
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>, Jagoda Jablonska <\email{jagoda100jablonska@gmail.com}>
}
\seealso{
\code{\link[cgmisc]{get.overlapping.windows}}, \code{\link[cgmisc]{het.for.overlap.wind}}
}
\keyword{heterozygosity,}
\keyword{window}

