% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.pac.r
\name{plot.pac}
\alias{plot.pac}
\title{Plot p-values of allele count differences between populations.}
\usage{
plot.pac(data, allele.cnt, plot.LD = F, legend.pos = "default", ...)
}
\arguments{
\item{data}{an object of the \code{\link[GenABEL]{gwaa.data-class}},}

\item{allele.cnt}{an object of the \code{\link[pac.result]{pac.result-class}} class,}

\item{plot.LD}{a logical indicating whether LD pattern should be plotted on top of p-values (experimental),}

\item{legend.pos}{position of a legend. If null, no legend is shown. If 'default', the default settings are used,}

\item{...}{}
}
\value{
null
}
\description{
Plots p-values of allele count differences between populations.
}
\details{
Plots p-values of allele count differences between populations.
}
\examples{
\dontrun{
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{allele}
\keyword{counts,}
\keyword{plot,}
\keyword{populations}

