% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.qq}
\alias{plot.qq}
\title{Plotting a QQ-plot with desired confidence intervals.}
\usage{
\method{plot}{qq}(pvals, conf = c(0.05, 0.95), conf.type = "shade",
  conf.col = rgb(0.5, 0.5, 1, 0.5), step = 1)
}
\arguments{
\item{pvals}{a vector of empirical p-values}

\item{conf}{a vector of the two thresholds for confidence interval}

\item{conf.type}{type of confidence interval plotting, conf.type is either "shade" or "line"}

\item{conf.col}{color for plotting confidence interval}

\item{step}{computation coarseness step}
}
\description{
Plots a QQ-plot with -log10(p-value) on both axes
and displays computed confidence intervals.
}
\details{
Calculation of confidence intervals:
The j-th order statistic from a uniform(0,1) sample has a beta(j,n-j+1) distribution.
For details see Casella & Berger.

step goes from 1 up. Step=1 - slow and fine determination of confidence interval.
Step=10 - fast and coarse determination of the confidence interval.
}
\examples{
\dontrun{
 an0 <- qtscore(bt~antibodyLevel,data.orig)
 pvals <- an0[,"P1df"]
 plot.qq(pvals, conf=c(0.05, 0.95), conf.type="lines", conf.col="tomato", step=10)
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\references{
Casella & Berger, 2002, 2nd edition, pg 230, Duxbury.
}
\keyword{bed,}
\keyword{gwas,}
\keyword{p-values}

