% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pop.allele.counts.r
\name{pop.allele.counts}
\alias{pop.allele.counts}
\title{Compare allele counts in two populations.}
\usage{
pop.allele.counts(data, pops, progress = TRUE)
}
\arguments{
\item{data}{a \code{\link[GenABEL]{gwaa.data-class}} object}

\item{pops}{a vector defining populations. The vector should be as long as
the number of individuals in data and populations are coded as 1 and 2.}
}
\value{
an object of \code{\link[cgmisc]{pac.result-class}}
}
\description{
Compare allele counts in two populations using Fisher exact test.
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}

