% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qq.emp}
\alias{qq.emp}
\title{QQ plot with empirical p-values}
\usage{
qq.emp(obs, conf = c(0.05, 0.95), result, step = 10, legend = T,
  plot.emp = T, conf.level = 0.95)
}
\arguments{
\item{obs}{observed values from association test}

\item{step}{computation coarseness step}

\item{results}{vector of empirical p-values}

\item{legend}{if the legend is to be plotted}

\item{plot.emp}{plotting empirical values}

\item{conf.level}{confidence level (0.95 by default)}
}
\description{
Plots qq plot with empirical values and empirical confidence intervals
}
\examples{
\dontrun{
 qq.emp(obs=data.mm[,'Pc1df], conf=c(0.05, 0.95), result = result[,'Pc1df'],step=10, legend=T, plot.emp=T, conf.level=.95)
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{emp}
\keyword{permutations}

