% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str-utils.R
\name{verybasename}
\alias{verybasename}
\title{Manipulate file paths}
\usage{
verybasename(path)
}
\arguments{
\item{path}{character vector, containing path names.}
}
\description{
Like \code{basename}, but removes also the file extension (i.e. everything
after the first dot in a file name).
}
\examples{
verybasename(c("/test1/my_file.txt.bck", "/test2/another_file.txt"))

}
