% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{vlookup function}
\usage{
vlookup(x, table, col_return, ..., col_search = 1)
}
\arguments{
\item{x}{Search criterion.}

\item{table}{Array.}

\item{col_return}{Index.}

\item{near_val}{Not yet implemented.}

\item{sort_order}{Not yet implemented.}
}
\description{
To do.
}
\examples{
shopping <- data.frame(product  = c("Sports Almanac", "Frisbee",
                                    "Water bottles"),
                       quantity = c(1, 3, 6))
seller <- data.frame(product  = c("Frisbee", "Lava lamp", "Sports Almanac",
                                  "Video camera", "Water bottles"),
                     price    = c(7, 20, 35, 75, 3))
shopping$unit_price <- vlookup(shopping$product, seller, "price")
shopping$total_price <- with(shopping, unit_price * quantity)
seller
shopping
sum(shopping$total_price)

}
