% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boxplot.snp.r
\name{boxplot.snp}
\alias{boxplot.snp}
\title{Box plot with genotypes}
\usage{
boxplot.snp(data, marker, trait, recode = F, ...)
}
\arguments{
\item{marker}{SNP name or index}

\item{trait}{trait name}

\item{recode}{if genotypes are to be recoded to 'AA','Aa','aa'.}

\item{GWASdata}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}}
}
\value{
null
}
\description{
Makes boxplots of the individuals in every genotypic class, based on a SNP.
Works for both outbreed (three boxes) and inbreed (two boxes) data.
}
\author{
Simon Forsberg <\email{simon.forsberg@slu.se}>
}

