% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute.fstats}
\alias{compute.fstats}
\title{Compute Fst fixation index for two populations}
\usage{
compute.fstats(data, pops)
}
\arguments{
\item{data}{a gwaa.data class object as used by \code{\link[GenABEL]{gwaa.data-class}}}

\item{pops}{a vector of two values indicating to which population an individual belongs}
}
\value{
an \code{\link[cgmisc]{fstats.result}} class object
}
\description{
Fixation index Fst is a measure of population differentiation due to genetic structure.
Given a set of genotypes in two populations, the function computes fixation index
Fst and corresponding indices: Fit and Fis.
}
\details{
\itemize{
  \item data -- a standard \code{\link[GenABEL]{gwaa.data-class}} object
  \item pops -- a vector of two values indicating to which population an individual belongs.
  Typically, one uses a vector of zeroes and ones where 0 marks an individual belonging to
  population 1 and 1 marks an individual belonging to population 2. Often, the vector is a result
  of clustering in MDS-scaled genomic kinship space
}
}
\examples{
\dontrun{
fstats  <- compute.Fstats(data, pops)
}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@slu.se}>
}
\references{
Holsinger, Kent E.; Bruce S. Weir (2009). "Genetics in geographically structured populations:
defining, estimating and interpreting FST". Nat Rev Genet 10 (9): 639-650. doi:10.1038/nrg2611. ISSN 1471-0056. PMID 19687804.
}
\seealso{
\code{\link[GenABEL]{gwaa.data-class}}, \code{\link[cgmisc]{fstats.result}}
}
\keyword{FST}
\keyword{fixation}
\keyword{heterozygosity}
\keyword{index}
\keyword{population}
\keyword{structure}

