% Generated by roxygen2 (4.0.2): do not edit by hand
\name{epistasis.scan}
\alias{epistasis.scan}
\title{epistasisScan}
\usage{
epistasis.scan(GWASdata, SNP, trait.name = NULL, pheno = NULL)
}
\arguments{
\item{GWASdata}{Object of class gwaa.data}

\item{SNP}{The name or index of one marker/SNP in GWASdata}

\item{trait.name}{The name of a phenotype in GWASdata}

\item{pheno}{A vector with phenotypes}
}
\value{
a list with all the models (lm-objects),
the p-values of every model and the p-values of every coefficient in every model.
}
\description{
Scans for epistasis between a specified SNP and all others.
}
\author{
Simon Forsberg  <\email{simon.forsberg@slu.se}>
}

