% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Format real numbers for pretty printing.}
\usage{
fmt(digits = 3, lh_equal = TRUE)
}
\arguments{
\item{digits}{How many digits to be kept?}

\item{lh_equ}{With leaf-hand side equal (or unequal) sign?}
}
\description{
Closure to format numeric numbers for pretty printing.
}
