\name{fstats.result}
\alias{fstats.result}
\alias{fstats.result-class}
\title{The \code{fstats.result} class}
\description{
  This class contains results of running the
  \code{\link[cgmisc]{compute.Fstats}} function.
}
\details{
  This line and the next ones go into the details. This
  line thus appears in the details as well.
}
\section{Slots}{
  \describe{ \item{\code{pops}:}{A \code{list} of
  per-population statistics} \item{\code{glob}:}{A
  \code{data.frame} of global statistics} }
}
\author{
  Marcin Kierczak <\email{Marcin.Kierczak@slu.se}>
}

