% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.clumps.r
\name{plot.clumps}
\alias{plot.clumps}
\title{Plot marker clumps on Manhattan plot.}
\usage{
plot.clumps(gwas.result, clumps, chr, region, clambda = F)
}
\arguments{
\item{gwas.result}{an object of the \code{\link[GenABEL]{gwaa.data-class}},}

\item{clumps}{a result of running the \code{\link[cgmisc]{clump.markers}} function,}

\item{chr}{chromosome to display,}

\item{region}{a vector of start and stop coordinates of a region to display,}

\item{clambda}{a logical indicating whether corrected Pc1df p-values are
to be used.}
}
\description{
Plot clumps resulting from running the \code{\link[cgmisc]{clump.markers}} function.
}
\examples{
\dontrun{plot.clumps(data, myclumps, 1, c(14172, 19239))}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\seealso{
\code{\link[cgmisc]{clump.markers}}
}
\keyword{clumping}
\keyword{clumps}
\keyword{plot}

