% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.fstats.r
\name{plot.fstats}
\alias{plot.Fst}
\alias{plot.fstats}
\title{Plot results of \code{compute.fstats}}
\usage{
plot.fstats(data, fstats, est.type = "Hudson", ...)
}
\arguments{
\item{data}{an object of the \code{\link[GenABEL]{gwaa.data-class}}}

\item{fstats}{an object of the \code{\link[cgmisc]{fstats.result-class}}}

\item{est.type}{Fst estimate to be plotted: 'WC' (Weir and Cockerham's), 'Nei', 'Hudson' (default) or
'naive' (based on Sewall Wright's original FST).}
}
\description{
Plot results obtained using \code{\link[cgmisc]{compute.fstats}} function.
}
\examples{
\dontrun{plot.fstats(data, fst)}
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\seealso{
\code{\link[cgmisc]{compute.fstats}}, \code{\link[cgmisc]{fstats.result-class}}
}
\keyword{FST,}
\keyword{Fst,}
\keyword{fst,}
\keyword{fstats,}
\keyword{plot}
\keyword{populations,}
\keyword{subpopulations}

