% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.pac}
\alias{plot.pac}
\title{Plot p-values of allele count differences between populations.}
\usage{
\method{plot}{pac}(data, allele.cnt, plot.LD = FALSE, ...)
}
\arguments{
\item{data}{an object of the \code{\link[GenABEL]{gwaa.data-class}},}

\item{allele.cnt}{an object of the \code{\link[pac.result]{pac.result-class}} class,}

\item{plot.LD}{a logical indicating whether LD pattern should be plotted on top of p-values (experimental),}

\item{...}{}
}
\value{
null
}
\description{
Plots p-values of allele count differences between populations.
}
\details{
Plots p-values of allele count differences between populations.
}
\examples{
\dontrun{
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}
\keyword{allele}
\keyword{counts,}
\keyword{plot,}
\keyword{populations}

