% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.qq.r
\name{plot.qq}
\alias{plot.qq}
\title{QQ plot with empirical p-values}
\usage{
plot.qq(data = NULL, obs, emp, N = 30, step = 10, legend = T,
  plot.emp = T, conf.level = 0.95, conf = c(0.025, 0.975), show.pb = T)
}
\arguments{
\item{data}{a gwaa.data-class object used to fit the original model}

\item{obs}{observed values from association test}

\item{emp}{a matrix of empirical p-values (permutation per row) or a list, result of running polygenic model}

\item{N}{if polygenic model supplied, a number of permutations to run}

\item{step}{computation coarseness step}

\item{legend}{if the legend is to be plotted}

\item{plot.emp}{plotting empirical values}

\item{conf.level}{confidence threshold (0.95 by default)}

\item{conf}{a vector defininy the lower and the upper confidence interval, default 5% CI.}

\item{show.pb}{a logical indicating whether progress bar will be shown}
}
\description{
Plots qq plot with empirical values and empirical confidence intervals
}
\examples{
\dontrun{
 plot.qq(obs=data.mm[,'Pc1df], conf=c(0.025, 0.975), emp = result[,'Pc1df'], step=10, legend=T, plot.emp=T, conf.level=.95)
 }
}
\author{
Marcin Kierczak <\email{Marcin.Kierczak@imbim.uu.se}>
}

