% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{vlookup function}
\usage{
vlookup(x, table, col_return, ..., col_search = 1)
}
\arguments{
\item{x}{Search criterion.}

\item{table}{Array.}

\item{col_return}{Index.}

\item{near_val}{Not yet implemented.}

\item{sort_order}{Not yet implemented.}
}
\description{
To do.
}
